requirements.txt
discord.py
firebase-admin
sqlite3
